// src/app/_core/models/models.ts
export interface RequestAnswersVO {
  id: number;
  idRequest: number;
  answer: string;
  type : string;
  label : string;
  idFormField : string;
}
// ...
export interface WelcomeMailVO {
  id: number;
  addressId: number;
  issueDate: Date;
  fullInfosWelcomeMail: string;
  civility: string;
  coach: string;
  fullTitle: string;
  isArchived: boolean;
  isValid: boolean;
  requestId: number;
}