// src/app/_core/models/request-customer-vo.ts
import { EntityBase } from './entity-base';
import { TaskLight } from './task-customer-vo';
import { InteractionLight } from './interaction-vo';
import { ReferenceData } from './reference-data';
export interface RequestCustomerVO extends EntityBase {
  customerId: number;
  idRequest: number;
  title: string;
  createdAt: Date;
  statuts: string;
  tasks: TaskLight[];
  task: TaskLight;
  interactions: InteractionLight[];
  description: string;
  requestTypeLabel: string;
  technicalKey: string;
  modifier: Date;
  firstNameInterlocuteur: String;
  lastNameInterlocuteur: String;
  idPorteur: number;
  firstNamePorteur: string;
  lastNamePorteur: string;
  firstNameDestinataire: String;
  lastNameDestinataire: String;
  idRespCom: number;
  firstNameRespCommerciale: String;
  lastNameRespCommerciale: String;
  idRespAffaire: number;
  firstNameRespAffaire: String;
  lastNameRespAffaire: String;
  emailRespAffaire: string;
  nodeAnswer: String;
  universLabel: string;
  mediaLabel: string;
  instanceId: string;
  closureComment: string;
  conclusionType: string;
  realEndDate: Date;
  universe: ReferenceData;
  labelUnivers: string;
  universId: number;
  requestTypeId: number;
  destinataireId: number;
  interlocutorId: number;
  isConnectedOnWf : boolean;
  arrow: string;
  cartStatus: string;
  cartId: number;
  cartColor: string;
  cartItemLentgh : number;
  workflowId : number; // Cet attribut a été ajouté pour pouvoir différencier entre les demandes manuelles d'Athena et de Chorniche
}